requirements.txt
beautifulsoup4==4.12.2
duckduckgo_search==4.1.1
fastapi==0.106.0
langchain==0.0.352
numpy==1.26.2
openai==1.6.1
requests==2.31.0
starlette==0.27.0
uvicorn==0.25.0
python-dotenv==1.0.0
sse_starlette
streamlit==1.29.0
